fix: service webhook should only validate FLB enabled service
